Moved blast material conversion to editor module.
