+++ Gems/Blast/Code/Source/Components/BlastSystemComponent.cpp
@@ -45,7 +45,6 @@ namespace Blast
 
         MaterialConfiguration::Reflect(context);
         MaterialAsset::Reflect(context);
-        ReflectLegacyMaterialClasses(context);
 
         if (AZ::SerializeContext* serialize = azrtti_cast<AZ::SerializeContext*>(context))
         {
